requirements.txt
asgiref==3.5.1
Django==4.0.4
django-geojson==3.2.1
djangorestframework==3.13.1
psycopg2==2.9.3
pytz==2022.1
sqlparse==0.4.2
